Checking for date changes, etc

+++ app/src/main/java/com/neirno/flower_jc_k/feature_flower/domain/repository/FlowerRepository.kt
@@ -7,7 +7,7 @@ interface FlowerRepository {
 
     fun getFlowers(): Flow<List<Flower>>
 
-    suspend fun getFlowerById(id: Int): Flower?
+    suspend fun getFlowerById(id: Long): Flower?
 
     suspend fun insertFlower(flower: Flower): Long
