+++ app/src/main/java/com/neirno/flower_jc_k/feature_flower/domain/use_case/GetFlower.kt
@@ -7,7 +7,7 @@ class GetFlower(
     private val repository: FlowerRepository
 ) {
 
-    suspend operator fun invoke(id: Int): Flower? {
+    suspend operator fun invoke(id: Long): Flower? {
         return repository.getFlowerById(id)
     }
 }
